# openapi.yaml
paths:
...
  /api/studio/sources:
    post:
      summary: Upload file, submit URL, paste text, or import git
      description: Upload file, submit URL, paste text, or import git repository.
      responses:
        '200':
          description: Success
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceJsonBody'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateSourceFileBody'
    get:
      summary: List all sources.
      description: List all sources.
      responses:
        '200':
          description: Success
  /api/studio/sources/{source_id}:
    get:
      summary: Get a source with cleaned text.
      description: Get a source with cleaned text.
      responses:
        '200':
          description: Success
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
    put:
      summary: Update source title or cleaned_text.
      description: Update source title or cleaned_text.
      responses:
        '200':
          description: Success
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSourceBody'
    delete:
      summary: Delete source and associated episodes/audio.
      description: Delete source and associated episodes/audio.
      responses:
        '200':
          description: Success
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/sources/{source_id}/cover:
    get:
      summary: Get cover art for a source.
      description: Get cover art for a source.
      responses:
        '200':
          description: Success
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Upload cover art for a source.
      description: Upload cover art for a source.
      responses:
        '200':
          description: Success
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SourceCoverUploadBody'
  /api/studio/sources/{source_id}/re-clean:
    post:
      summary: Re-run normalizer on a source with different optio
      description: Re-run normalizer on a source with different options.
      responses:
        '200':
          description: Success
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReCleanSourceBody'
  /api/studio/sources/{source_id}/move:
    put:
      summary: Move a source to a folder.
      description: Move a source to a folder.
      responses:
        '200':
          description: Success
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveToFolderBody'
...
  /api/studio/folders:
    post:
      summary: Create a new folder.
      description: Create a new folder.
      responses:
        '200':
          description: Success
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderBody'
  /api/studio/folders/{folder_id}:
    put:
      summary: Rename or move a folder.
      description: Rename or move a folder.
      responses:
        '200':
          description: Success
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFolderBody'
    delete:
      summary: Delete a folder and all its contents (episodes aud
      description: Delete a folder and all its contents (episodes audio files are
        deleted).
      responses:
        '200':
          description: Success
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/folders/{folder_id}/playlist:
    post:
      summary: Start playing all episodes in a folder as a playli
      description: Start playing all episodes in a folder as a playlist.
      responses:
        '200':
          description: Success
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/folders/{folder_id}/episodes:
    get:
      summary: Get all episodes in a folder for playlist building
      description: Get all episodes in a folder for playlist building.
      responses:
        '200':
          description: Success
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/reorder:
    post:
      summary: Batch update sort orders.
      description: Batch update sort orders.
      responses:
        '200':
          description: Success
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderBody'
...
info:
  title: OpenVox API
  version: 1.0.0
openapi: 3.0.3
components:
  schemas:
...
    CleaningSettings:
      type: object
      properties:
        code_block_rule:
          type: string
          description: 'How to handle code blocks: skip, read, summarize'
        remove_non_text:
          type: boolean
        handle_tables:
          type: boolean
        speak_urls:
          type: boolean
        expand_abbreviations:
          type: boolean
        preserve_parentheses:
          type: boolean
        preserve_structure:
          type: boolean
        paragraph_spacing:
          type: integer
        section_spacing:
          type: integer
        list_item_spacing:
          type: integer
      additionalProperties: false
    UrlSettings:
      type: object
      properties:
        use_jina:
          type: boolean
          description: Use Jina Reader for URL extraction
        jina_fallback:
          type: boolean
          description: Fall back to Jina if other methods fail
      additionalProperties: false
    CreateSourceJsonBody:
      type: object
      properties:
        text:
          type: string
          description: Raw text to import
        title:
          type: string
          description: Source title
        url:
          type: string
          description: URL to import content from
        git_url:
          type: string
          description: Git repository URL to import
        git_subpath:
          type: string
          description: Subdirectory path within the git repo
        cleaning_settings:
          description: Override cleaning options
          allOf:
          - $ref: '#/components/schemas/CleaningSettings'
        url_settings:
          description: URL extraction settings
          allOf:
          - $ref: '#/components/schemas/UrlSettings'
      additionalProperties: false
    CreateSourceFileBody:
      type: object
      properties:
        file:
          type: string
          format: binary
          description: File to upload (.md, .txt)
      additionalProperties: false
    SourceCoverUploadBody:
      type: object
      properties:
        cover:
          type: string
          format: binary
          description: Cover art image file
      additionalProperties: false
    UpdateSourceBody:
      type: object
      properties:
        title:
          type: string
        cleaned_text:
          type: string
      additionalProperties: false
    ReCleanSourceBody:
      type: object
      properties:
        code_block_rule:
          type: string
        remove_non_text:
          type: boolean
        handle_tables:
          type: boolean
        speak_urls:
          type: boolean
        expand_abbreviations:
          type: boolean
        preserve_parentheses:
          type: boolean
      additionalProperties: false
    MoveToFolderBody:
      type: object
      properties:
        folder_id:
          type: string
          description: Target folder ID, or null for root
          nullable: true
      additionalProperties: false
...
    CreateFolderBody:
      type: object
      properties:
        name:
          type: string
          default: New Folder
        parent_id:
          type: string
          nullable: true
        sort_order:
          type: integer
          default: 0
      additionalProperties: false
    UpdateFolderBody:
      type: object
      properties:
        name:
          type: string
        parent_id:
          type: string
          nullable: true
        sort_order:
          type: integer
      additionalProperties: false
    ReorderItem:
      type: object
      properties:
        type:
          type: string
        id:
          type: string
        sort_order:
          type: integer
      required:
      - id
      - sort_order
      additionalProperties: false
    ReorderBody:
      type: object
      properties:
        items:
          type: array
          default: []
          items:
            $ref: '#/components/schemas/ReorderItem'
      additionalProperties: false
...